(trunk) Oneiric compile fixes.
git-svn-id: http://svn.gplates.org/projects/gplates/trunk@12780 e74ed31e-5968-4cd8-ac0b-be5e0f3b170c

+++ src/file-io/MipmappedRasterFormatReader.h
@@ -1225,7 +1225,7 @@ namespace GPlatesFileIO
 
 // 				qDebug() << "expected_file_size " << expected_file_size;
 // 				qDebug() << "file_info.size() " << file_info.size();
-				if (expected_file_size != file_info.size())
+				if (expected_file_size != static_cast<quint64>(file_info.size()))
 				{
 					throw FileFormatNotSupportedException(
 							GPLATES_EXCEPTION_SOURCE, "detected a partially written mipmap file");
